refactoring in 'users' , 'chats', 'group_tags', 'common_methods' tests
