--- tests/test.py
@@ -1,3 +1,0 @@
-a = {}
-if a is None:
-    print('aaaa')
